replace korlibs kbignum with ionspin bignum

+++ web3/src/commonMain/kotlin/dev.icerock.moko.web3/contract/UInt256Param.kt
@@ -4,13 +4,14 @@
 
 package dev.icerock.moko.web3.contract
 
-import com.soywiz.kbignum.BigInt
+import com.ionspin.kotlin.bignum.integer.BigInteger
+import com.ionspin.kotlin.bignum.integer.toBigInteger
 import dev.icerock.moko.web3.hex.internal.hexStringToByteArray
 import dev.icerock.moko.web3.hex.internal.toHex
 
-object UInt256Param : StaticEncoder<BigInt> {
+object UInt256Param : StaticEncoder<BigInteger> {
     @OptIn(ExperimentalStdlibApi::class)
-    override fun encode(item: BigInt): ByteArray {
+    override fun encode(item: BigInteger): ByteArray {
         val dataByteArray = item.toString(16).hexStringToByteArray(unsafe = true)
 
         return when {
@@ -20,8 +21,8 @@ object UInt256Param : StaticEncoder<BigInt> {
     }
 
     @OptIn(ExperimentalStdlibApi::class)
-    override fun decode(source: ByteArray): BigInt {
+    override fun decode(source: ByteArray): BigInteger {
         val value = source.toHex()
-        return BigInt(value, 16)
+        return value.toBigInteger(16)
     }
 }
