+++ web3/src/commonMain/kotlin/dev.icerock.moko.web3/hex/Transformers.kt
@@ -4,7 +4,7 @@
 
 package dev.icerock.moko.web3.hex
 
-import com.soywiz.kbignum.BigInt
+import com.ionspin.kotlin.bignum.integer.BigInteger
 import dev.icerock.moko.web3.hex.internal.hexBytesFillToSizedHex
 import dev.icerock.moko.web3.hex.internal.hexStringFillToSizedHex
 
@@ -13,7 +13,7 @@ fun HexString.fillToStrict() = fillToSizedHex(
         override val size = this@fillToStrict.size
 
         override fun createInstance(value: String): HexString = HexString.Lenient.createInstance(value)
-        override fun createInstance(value: BigInt): HexString = HexString.Lenient.createInstance(value)
+        override fun createInstance(value: BigInteger): HexString = HexString.Lenient.createInstance(value)
         override fun createInstance(value: ByteArray): HexString = HexString.Lenient.createInstance(value)
     }
 )
