+++ web3/src/commonMain/kotlin/dev.icerock.moko.web3/requests/polling/NewBlocksPolling.kt
@@ -4,39 +4,23 @@
 
 package dev.icerock.moko.web3.requests.polling
 
-import com.soywiz.kbignum.BigInt
-import dev.icerock.moko.web3.BlockHash
+import com.ionspin.kotlin.bignum.integer.BigInteger
 import dev.icerock.moko.web3.BlockInfo
 import dev.icerock.moko.web3.BlockState
 import dev.icerock.moko.web3.Web3Executor
 import dev.icerock.moko.web3.requests.Web3Requests
 import dev.icerock.moko.web3.requests.getBlockNumber
-import kotlinx.coroutines.ExperimentalCoroutinesApi
-import kotlinx.coroutines.channels.Channel
-import kotlinx.coroutines.channels.produce
-import kotlinx.coroutines.coroutineScope
-import kotlinx.coroutines.currentCoroutineContext
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.SharingStarted
-import kotlinx.coroutines.flow.asFlow
-import kotlinx.coroutines.flow.channelFlow
-import kotlinx.coroutines.flow.collect
-import kotlinx.coroutines.flow.distinctUntilChanged
-import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.flow.flow
-import kotlinx.coroutines.flow.forEach
-import kotlinx.coroutines.flow.map
-import kotlinx.coroutines.flow.receiveAsFlow
-import kotlinx.coroutines.flow.shareIn
 import kotlinx.coroutines.flow.transform
 
 
-private operator fun BigInt.rangeTo(other: BigInt): Iterable<BigInt> =
+private operator fun BigInteger.rangeTo(other: BigInteger): Iterable<BigInteger> =
     generateSequence(seed = this) { block -> (block + 1).takeIf { it <= other} }.asIterable()
 
 fun Web3Executor.newBlocksShortPolling(
-    fromBlock: BigInt? = null,
+    fromBlock: BigInteger? = null,
     pollingInterval: Long = 5_000
 ): Flow<BlockInfo> =
     flow {
@@ -49,7 +33,7 @@ fun Web3Executor.newBlocksShortPolling(
                 previousBlockNumber = blockNumber
             }
         }
-    }.transform { (fromBlock, toBlock): Pair<BigInt, BigInt> ->
+    }.transform { (fromBlock, toBlock): Pair<BigInteger, BigInteger> ->
         val blockNumbers = fromBlock..(toBlock - 1)
 
         val requests = blockNumbers
